Move pipeline set types into pipeline::set
