+++ gotham/src/pipeline/mod.rs
@@ -1,6 +1,7 @@
 //! Defines types for a middleware pipeline
 
 pub mod chain;
+pub mod set;
 pub mod single;
 
 use std::io;
@@ -32,11 +33,12 @@ use state::{request_id, State};
 /// # use gotham::handler::HandlerFuture;
 /// # use gotham::middleware::{Middleware, NewMiddleware};
 /// # use gotham::pipeline::new_pipeline;
+/// # use gotham::pipeline::set::*;
 /// # use gotham::router::Router;
 /// # use gotham::router::tree::TreeBuilder;
 /// # use gotham::router::route::{RouteImpl, Extractors, Delegation};
 /// # use gotham::router::route::matcher::MethodOnlyRouteMatcher;
-/// # use gotham::router::route::dispatch::{new_pipeline_set, finalize_pipeline_set, DispatcherImpl};
+/// # use gotham::router::route::dispatch::DispatcherImpl;
 /// # use gotham::test::TestServer;
 /// # use gotham::extractor::{NoopPathExtractor, NoopQueryStringExtractor};
 /// # use gotham::router::response::finalizer::ResponseFinalizerBuilder;
