+++ gotham/src/router/builder/single.rs
@@ -24,7 +24,8 @@ use handler::{Handler, NewHandler};
 /// # use gotham::router::builder::*;
 /// # use gotham::pipeline::new_pipeline;
 /// # use gotham::middleware::session::NewSessionMiddleware;
-/// # use gotham::router::route::dispatch::{new_pipeline_set, finalize_pipeline_set};
+/// # use gotham::pipeline::set::*;
+/// #
 /// fn my_handler(_: State) -> (State, Response) {
 ///     // Handler implementation elided.
 /// #   unimplemented!()
@@ -61,8 +62,9 @@ pub trait DefineSingleRoute {
     /// # use gotham::router::Router;
     /// # use gotham::router::builder::*;
     /// # use gotham::pipeline::new_pipeline;
+    /// # use gotham::pipeline::set::*;
     /// # use gotham::middleware::session::NewSessionMiddleware;
-    /// # use gotham::router::route::dispatch::{new_pipeline_set, finalize_pipeline_set};
+    /// #
     /// fn my_handler(_: State) -> (State, Response) {
     ///     // Handler implementation elided.
     /// #   unimplemented!()
@@ -101,8 +103,9 @@ pub trait DefineSingleRoute {
     /// # use gotham::router::Router;
     /// # use gotham::router::builder::*;
     /// # use gotham::pipeline::new_pipeline;
+    /// # use gotham::pipeline::set::*;
     /// # use gotham::middleware::session::NewSessionMiddleware;
-    /// # use gotham::router::route::dispatch::{new_pipeline_set, finalize_pipeline_set};
+    /// #
     /// struct MyNewHandler;
     /// struct MyHandler;
     ///
@@ -157,8 +160,9 @@ pub trait DefineSingleRoute {
     /// # use gotham::router::Router;
     /// # use gotham::router::builder::*;
     /// # use gotham::pipeline::new_pipeline;
+    /// # use gotham::pipeline::set::*;
     /// # use gotham::middleware::session::NewSessionMiddleware;
-    /// # use gotham::router::route::dispatch::{new_pipeline_set, finalize_pipeline_set};
+    /// #
     /// #[derive(Deserialize, StateData, StaticResponseExtender)]
     /// struct MyPathParams {
     /// #   #[allow(dead_code)]
@@ -214,8 +218,8 @@ pub trait DefineSingleRoute {
     /// # use gotham::router::Router;
     /// # use gotham::router::builder::*;
     /// # use gotham::pipeline::new_pipeline;
+    /// # use gotham::pipeline::set::*;
     /// # use gotham::middleware::session::NewSessionMiddleware;
-    /// # use gotham::router::route::dispatch::{new_pipeline_set, finalize_pipeline_set};
     /// #
     /// #[derive(StateData, Deserialize, StaticResponseExtender)]
     /// struct MyQueryParams {
