+++ gotham/src/router/route/mod.rs
@@ -12,11 +12,11 @@ use std::panic::RefUnwindSafe;
 
 use hyper::{Response, Uri};
 
-use http::request::query_string;
-use router::route::dispatch::Dispatcher;
 use handler::HandlerFuture;
+use http::request::query_string;
 use extractor::{self, PathExtractor, QueryStringExtractor};
 use router::non_match::RouteNonMatch;
+use router::route::dispatch::Dispatcher;
 use router::route::matcher::RouteMatcher;
 use router::tree::SegmentMapping;
 use state::{request_id, State};
@@ -87,8 +87,9 @@ pub struct ExtractorFailed;
 /// #
 /// # use gotham::http::response::create_response;
 /// # use gotham::extractor::{NoopPathExtractor, NoopQueryStringExtractor};
+/// # use gotham::pipeline::set::*;
 /// # use gotham::router::route::matcher::MethodOnlyRouteMatcher;
-/// # use gotham::router::route::dispatch::{new_pipeline_set, finalize_pipeline_set, DispatcherImpl};
+/// # use gotham::router::route::dispatch::DispatcherImpl;
 /// # use gotham::state::State;
 /// # use gotham::router::route::{RouteImpl, Extractors, Delegation};
 /// #
@@ -117,8 +118,9 @@ pub struct ExtractorFailed;
 /// #
 /// # use gotham::http::response::create_response;
 /// # use gotham::extractor::{NoopPathExtractor, NoopQueryStringExtractor};
+/// # use gotham::pipeline::set::*;
 /// # use gotham::router::route::matcher::MethodOnlyRouteMatcher;
-/// # use gotham::router::route::dispatch::{new_pipeline_set, finalize_pipeline_set, DispatcherImpl};
+/// # use gotham::router::route::dispatch::DispatcherImpl;
 /// # use gotham::state::State;
 /// # use gotham::router::Router;
 /// # use gotham::router::route::{RouteImpl, Extractors, Delegation};
